tests: refactor approvaltests Namer
Move Namer to own module.

Approval files are now stored in `/tests/approvals` instead of
`/tests/integrationtests/approved`.
