+++ tests/approvals/integrationtests/api/test_indicators_atribute_completeness.py-test_indicators_attribute_completeness_filter_invalid[headers0-schema0].approved.txt
