+++ tests/approvals/integrationtests/api/test_indicators_attribute_completeness.py-test_indicators_attribute_completeness_filter_invalid[headers1-schema1].approved.txt
