+++ tests/approvals/integrationtests/indicators/test_attribute_completeness.py-TestCalculation-test_calculate[indicator0].approved.txt
