+++ tests/approvals/integrationtests/indicators/test_attribute_completeness.py-test_create_description_attribute_keys_single.approved.txt
