+++ tests/approvals/integrationtests/indicators/test_building_comparison.py-TestCalculate-test_calculate.approved.txt
