+++ tests/approvals/integrationtests/indicators/test_building_comparison.py-TestCalculate-test_calculate_above_one_th.approved.txt
