+++ tests/approvals/integrationtests/indicators/test_building_comparison.py-TestCalculate-test_calculate_some_intersection.approved.txt
