+++ tests/approvals/integrationtests/indicators/test_road_comparison.py-TestCalculate-test_calculate_no_intersection.approved.txt
