+++ tests/approvals/integrationtests/indicators/test_road_comparison.py-TestCalculate-test_calculate_reference_lenght_0.approved.txt
